Title: 요일 맞추기, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 250521/요일 맞추기/guess-day-of-week.java
@@ -31,6 +31,6 @@ public class Main {
         long cal = (1+totalDays) %7;
         String calWeek = week[cal];
 
-        System.out.println(calWeek);
+        System.out.println((int)calWeek);
     }
 }
